n3pro: sync file-backed netfm presets, paged web remote and local-only gating
Overlay + n3pro_port.patch update for the matching rockbox commit
(netfm reads its presets from a file with no fixed cap; the web remote
pages its file/playlist lists; web control is limited to local file
playback).  Also drops the Bluetooth PoC sources and BLUETOOTH.md.

+++ apps/netfm_stream.c
@@ -667,7 +667,7 @@ static void ts_emit_pes(const unsigned char *pes, size_t len)
              * it here (the probe runs before the codec starts) */
             int sfi = (pes[i + 2] >> 2) & 0xf;
             if (sfi < 13 && aac_samplerates[sfi] &&
-                source.id3.frequency != aac_samplerates[sfi])
+                source.id3.frequency != (unsigned long)aac_samplerates[sfi])
             {
                 source.id3.frequency = aac_samplerates[sfi];
                 pthread_mutex_lock(&source.lock);
